Add debian base,  deb 11 and 12 configs
Added an optional "base" key to the structure of the distro
configs.  When loaded by the config processor it will look for
the base key.  If it exists it will load the config file to which
that config points and merge the configs, otherwise it will
just use the named distro configs.